long term and link with scroll

+++ api/models/Long.js
@@ -1,12 +1,12 @@
 /**
- * Short.js
+ * Long.js
  *
  * @description :: TODO: You might write a short summary of how this model works and what it represents here.
  * @docs        :: http://sailsjs.org/documentation/concepts/models-and-orm/models
  */
 
 module.exports = {
-  tableName: 'short',
+  tableName: 'long',
   attributes: {
     partner: {
       type: 'string',
@@ -77,6 +77,5 @@ module.exports = {
       type: 'string',
       size: 255
     },
-  },
-};
+  },};
